hydra-cli:refactor introduce a codegen folder
